Body parameters are the parsed body.
Per conversations with @Crell, @ircmaxell, @pmjones, and several others, the
body parameters are always the parsed body content; calling them out as such,
and without the "parameters" verbiage, implies that parsing may create non-array
values, which is exceptionally likely when working with API and/or
non-form-encoded payloads.

Moreover, the content will always be specific to the incoming request, so
supporting one method representing POST and another representing parsed content
is redundant and potentially confusing.

Finally, considering the fact that the parsed body is an ambiguous domain,
forcing an ubiquitous signature is an unnecessary and limiting constraint.

However, IF the request method is POST, AND the Content-Type is
application/x-www-form-urlencoded, THEN the parsed body MUST be
equivalent to $_POST.